backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
firebase-admin==7.4.0
requests==2.33.1
pydantic==2.13.2
python-multipart==0.0.26
python-dotenv==1.1.0
